+++ src/main/java/com/hanghae/cloneinstagram/rest/like/dto/LikePostUserInterface.java
@@ -2,6 +2,7 @@ package com.hanghae.cloneinstagram.rest.like.dto;
 
 public interface LikePostUserInterface {
     String getProfile_url();
+    Long getUser_id();
     String getUsername();
     Long getFollow();
 }
