+++ src/main/java/com/hanghae/cloneinstagram/rest/post/dto/PostUsernameInterface.java
@@ -5,6 +5,7 @@ import java.time.LocalDateTime;
 public interface PostUsernameInterface {
      Long getId();
      String getProfile_url();
+     Long getUser_id();
      String getContent();
      String getImg_url();
      String getUsername();
